+++ CHANGELOG.md
@@ -11,11 +11,14 @@ This file documents the changes made to the formatter with each release.
 ### Changed
 
 - Force short lines that were manually wrapped on multiple lines with commas to merge back into one line when they fit within the max line length (this behavior was a leftover from the old formatter implementation)
+- Changed editor settings for safe and reorder mode; they're now a single format mode
+- Deprecated `--safe` flag, renamed into `--verify-structure`
 
 ### Fixed
 
 - Fix type casts with type subscripts wrapping when placed at the end of long lines (e.g. `[long, array] as Array[SomeType]`)
 - Fix icon annotation being reordered below class_name when extends was before class_name (#295)
+- Godot addon: Fix error when running the formatter with both safe and reorder mode active; they're now mutually exclusive in the editor settings (#286)
 
 ## Release 0.22.2 (2026-07-22)